feat: consolidate status color definitions and enhance workflow stage handling in CICD components
