+++ src/Shared/Components/CICDHistory/index.tsx
@@ -30,3 +30,4 @@ export { default as LogsRenderer } from './LogsRenderer'
 export * from './DeploymentHistoryConfigDiff'
 export * from './CiPipelineSourceConfig'
 export * from './StatusFilterButtonComponent'
+export { statusColor } from './constants'
